fix(suite): Code review updates
