+++ packages/components/src/components/Passphrase/consts.ts
@@ -1,0 +1,1 @@
+export const DOT = '●';
