+++ packages/components/src/components/Passphrase/types.ts
@@ -1,0 +1,1 @@
+export type WalletType = 'standard' | 'hidden';
