+++ packages/suite/src/components/suite/layouts/SuiteLayout/SuiteLayout.tsx
@@ -19,7 +19,7 @@ import { Sidebar } from './Sidebar/Sidebar';
 import { CoinjoinBars } from './CoinjoinBars/CoinjoinBars';
 import { MobileAccountsMenu } from 'src/components/wallet/WalletLayout/AccountsMenu/MobileAccountsMenu';
 import { selectSelectedAccount } from 'src/reducers/wallet/selectedAccountReducer';
-import { useAppShortcuts } from './utils';
+import { useAppShortcuts } from './useAppShortcuts';
 
 export const SCROLL_WRAPPER_ID = 'layout-scroll';
 export const Wrapper = styled.div`
