+++ packages/suite/src/components/suite/modals/ReduxModal/DeviceContextModal/types.ts
@@ -1,0 +1,1 @@
+export type ContentType = 'step1' | 'step2' | 'step3';
